Data generation

A tool to generate the dependent variable, with fixed covariates.

The dependent variable could be generate for more than 100 distributions which have at maximum 4 parameters. Each parameter could be defined in terms of covariates and random-effects.

The dependent variable, when normal, could include both an additional error-term and heteroschedasticity.

data_gen(
  Ni,
  M,
  D,
  k,
  b1,
  x1,
  b2 = NULL,
  x2 = NULL,
  b3 = NULL,
  x3 = NULL,
  b4 = NULL,
  x4 = NULL,
  xh = NULL,
  Dis,
  l,
  sigma = NULL,
  sigmah = NULL,
  sigmae = NULL,
  ty = NULL,
  costh = NULL,
  seed = NULL,
  id = NULL
)

Arguments

Ni

1xD vector containing the number of units (in population) for each area

M

Number of replicates

D

Number of areas

k

A vector, of maximum length equal to 4, of constant (intercept) used to generate data, i.e. mu=k+b1*x

b1

A vector of regression coefficients used for mu

x1

A matrix or a dataset where each column is a covariate used for mu

b2

A vector of regression coefficients used for sigma

x2

A matrix or a dataset where each column is a covariate used for sigma

b3

A vector of regression coefficients used for nu

x3

A matrix or a dataset where each column is a covariate used for nu

b4

A vector of regression coefficients used for tau

x4

A matrix or a dataset where each column is a covariate used for tau

xh

A vector of covariate used to generate heteroschedasticity

Dis

Type of distribution in form of rDis where Dis is one of the distribution allowe by GAMLSS. It is to be defined to use distribution different from the nornal one

l

A vector, of maximum length equal to 4, in which each space is a function (the inverse of the link-function) to be applied to the corresponding paramter, i.e. mu, sigma, nu and tau.

sigma

A vector, of maximum length equal to 4, of standard deviation used to generate random effects

sigmah

The standard deviation used to generate heteroschedasticity

sigmae

The standard deviation used to generate the error term when the link-function is the identity one

ty

To be specified equal to "no" if the generation process is mu=k+b1*x+e or equal to "het" if the generation process have to include heteroschedasticity

costh

A costant used to generate heteroschedasticity

seed

The seed to be used

id

Unit id.

Value

A list of data frame. Each one has: the dependent variable y, a set of fixed covariates, the small area column (sa) and the id column (id)

Note

The definition of the heteroschedastic term follows Ramirez-Aldana, R., & Naranjo, L. (2021).

References

Mori, L., & Ferrante, M. R. (2023). Small area estimation under unit-level generalized additive models for location, scale and shape. arXiv e-prints, arXiv-2302.

Ramirez-Aldana, R., & Naranjo, L. (2021). Random intercept and linear mixed models including heteroscedasticity in a logarithmic scale: Correction terms and prediction in the original scale. PloS one, 16(4), e0249910.

Rigby, R. A., & Stasinopoulos, D. M. (2005). Generalized additive models for location, scale and shape. Journal of the Royal Statistical Society: Series C (Applied Statistics), 54(3), 507–554.

Author

Lorenzo Mori and Maria Rosaria Ferrante

Examples

#Normal data
data_gen(Ni=rep(10,4), D=4, M=2 ,ty ="no", k=100, b1=4,
         x1=rnorm(40, 0,1),  b2 = NULL, x2 = NULL, b3 = NULL,
         x3 = NULL, b4 = NULL, x4 = NULL, xh = NULL,
         Dis, l=c(identity), sigma = 6,sigmah = NULL,
         sigmae = 22,  costh = NULL, seed = 1234)
#> Generating population 1 of 2
#> Generating population 2 of 2
#> [[1]]
#>            y          x1 sa id
#> 1  111.63653  0.85523991  1  1
#> 2  108.32712 -2.10178861  1  2
#> 3  114.48756 -0.68925864  1  3
#> 4   86.30673  0.46067181  1  4
#> 5   90.26009  0.03792886  1  5
#> 6  101.50067  0.38024644  1  6
#> 7  123.78570  2.24141716  1  7
#> 8  124.74209 -1.38194497  1  8
#> 9   92.00609  0.32687173  1  9
#> 10 117.72362  0.85859204  1 10
#> 11 110.55234 -0.02214594  2 11
#> 12  89.99900 -0.21938651  2 12
#> 13 116.71092 -0.56320735  2 13
#> 14 100.00259  0.87769493  2 14
#> 15  77.64034  0.18094721  2 15
#> 16  72.67831  0.19996814  2 16
#> 17 112.32878  1.14432751  2 17
#> 18  84.71127 -1.73917608  2 18
#> 19  82.21766 -0.54181966  2 19
#> 20 136.07150  1.03524821  2 20
#> 21  92.79158  1.81537970  3 21
#> 22  96.83571  1.53852422  3 22
#> 23 149.75733  0.43652414  3 23
#> 24 106.08738 -1.49933089  3 24
#> 25  89.05283 -0.47906499  3 25
#> 26 113.84498 -1.18931480  3 26
#> 27 120.64953  0.10356889  3 27
#> 28 101.13740  0.16556390  3 28
#> 29  99.20857 -1.59125949  3 29
#> 30 114.26201 -0.62094229  3 30
#> 31  78.19771 -0.05658567  4 31
#> 32 107.68798  0.38129030  4 32
#> 33  52.21091  0.25301100  4 33
#> 34 111.41196 -0.62717151  4 34
#> 35  72.31295 -0.30530441  4 35
#> 36 113.88083  0.65055626  4 36
#> 37  82.47104 -0.43946510  4 37
#> 38  81.23973  0.38282524  4 38
#> 39  99.71955  0.73221617  4 39
#> 40 115.42238 -2.00016320  4 40
#> 
#> [[2]]
#>            y          x1 sa id
#> 1  110.75703  0.85523991  1  1
#> 2  112.70160 -2.10178861  1  2
#> 3   63.86549 -0.68925864  1  3
#> 4   72.91713  0.46067181  1  4
#> 5  116.53586  0.03792886  1  5
#> 6   93.59026  0.38024644  1  6
#> 7  164.33547  2.24141716  1  7
#> 8   84.85513 -1.38194497  1  8
#> 9   93.09595  0.32687173  1  9
#> 10  87.14103  0.85859204  1 10
#> 11 109.95838 -0.02214594  2 11
#> 12  93.65919 -0.21938651  2 12
#> 13  92.43728 -0.56320735  2 13
#> 14 109.13070  0.87769493  2 14
#> 15 116.89142  0.18094721  2 15
#> 16  71.87128  0.19996814  2 16
#> 17  83.54042  1.14432751  2 17
#> 18  95.48782 -1.73917608  2 18
#> 19  92.55882 -0.54181966  2 19
#> 20  88.07124  1.03524821  2 20
#> 21 131.41605  1.81537970  3 21
#> 22 104.76720  1.53852422  3 22
#> 23 148.06363  0.43652414  3 23
#> 24  84.71634 -1.49933089  3 24
#> 25  58.80417 -0.47906499  3 25
#> 26  99.06289 -1.18931480  3 26
#> 27  79.64057  0.10356889  3 27
#> 28 106.77697  0.16556390  3 28
#> 29  88.95384 -1.59125949  3 29
#> 30  73.56449 -0.62094229  3 30
#> 31  67.39195 -0.05658567  4 31
#> 32  88.77531  0.38129030  4 32
#> 33  75.10179  0.25301100  4 33
#> 34  93.97151 -0.62717151  4 34
#> 35 149.57370 -0.30530441  4 35
#> 36  98.64602  0.65055626  4 36
#> 37 105.18741 -0.43946510  4 37
#> 38  97.80311  0.38282524  4 38
#> 39 141.42454  0.73221617  4 39
#> 40  88.46635 -2.00016320  4 40
#> 
#
#Heteroschedastic Normal data
#
data_gen(Ni=rep(10,4), D=4, M=2 ,ty ="het", k=100, b1=4,
         x1=rnorm(40, 0,1), b2 = NULL, x2 = NULL, b3 = NULL,
         x3 = NULL, b4 = NULL, x4 = NULL, xh = rnorm(40, 0,1),
         Dis, l=c(identity), sigma = 6, sigmah = 1, sigmae = 22,
         costh = 0.1, seed = 1234)
#> Generating population 1 of 2
#> Generating population 2 of 2
#> [[1]]
#>            y          x1 sa id
#> 1  102.26951  0.44151352  1  1
#> 2   87.26521  0.49075674  1  2
#> 3   64.32065  0.83686256  1  3
#> 4   80.79872  0.84365761  1  4
#> 5   47.47689  0.11396042  1  5
#> 6  132.04608  0.23383507  1  6
#> 7  204.82402 -0.33374960  1  7
#> 8   47.68943 -0.72348666  1  8
#> 9   89.66685  1.19740623  1  9
#> 10 108.93305 -0.97213617  1 10
#> 11 107.41227 -0.36185077  2 11
#> 12 104.57132  0.72577517  2 12
#> 13  99.47918  0.56551336  2 13
#> 14 112.67618  0.13956527  2 14
#> 15  96.12969 -0.91253419  2 15
#> 16 118.37301  0.92444445  2 16
#> 17 108.25327 -0.53238221  2 17
#> 18  91.36636  0.56246045  2 18
#> 19  91.17417  0.25234244  2 19
#> 20 105.94973  2.06930277  2 20
#> 21  89.05134 -0.68336832  3 21
#> 22 102.07188  1.47672180  3 22
#> 23  92.56853 -0.98954962  3 23
#> 24 102.35284  0.13893230  3 24
#> 25  85.35429  0.29615067  3 25
#> 26  81.65411 -1.49428831  3 26
#> 27  96.48810  0.92666633  3 27
#> 28  98.60799 -0.44733557  3 28
#> 29  90.16208 -2.24857604  3 29
#> 30  92.60279 -1.30668702  3 30
#> 31 113.45718  0.96292158  4 31
#> 32 101.13560 -0.26999602  4 32
#> 33  98.98758  0.70468768  4 33
#> 34  39.26999 -0.94204826  4 34
#> 35 109.60780 -1.25829569  4 35
#> 36  97.14837  0.76671910  4 36
#> 37  89.35832 -0.03750549  4 37
#> 38 119.65912 -0.04034886  4 38
#> 39 127.20947  1.40282663  4 39
#> 40 115.96748  0.49841992  4 40
#> 
#> [[2]]
#>            y          x1 sa id
#> 1   85.68939  0.44151352  1  1
#> 2  113.83138  0.49075674  1  2
#> 3  100.72342  0.83686256  1  3
#> 4   77.39220  0.84365761  1  4
#> 5   89.12411  0.11396042  1  5
#> 6  106.55112  0.23383507  1  6
#> 7  106.86512 -0.33374960  1  7
#> 8   94.55619 -0.72348666  1  8
#> 9   89.82026  1.19740623  1  9
#> 10  90.25380 -0.97213617  1 10
#> 11  87.30009 -0.36185077  2 11
#> 12 112.58682  0.72577517  2 12
#> 13  74.21238  0.56551336  2 13
#> 14 105.41559  0.13956527  2 14
#> 15  88.90338 -0.91253419  2 15
#> 16  98.59551  0.92444445  2 16
#> 17  88.74478 -0.53238221  2 17
#> 18 113.54720  0.56246045  2 18
#> 19  96.76452  0.25234244  2 19
#> 20 114.89082  2.06930277  2 20
#> 21  97.65410 -0.68336832  3 21
#> 22  82.75531  1.47672180  3 22
#> 23 134.94977 -0.98954962  3 23
#> 24 130.03186  0.13893230  3 24
#> 25  60.89238  0.29615067  3 25
#> 26  90.26427 -1.49428831  3 26
#> 27  96.14884  0.92666633  3 27
#> 28  61.70575 -0.44733557  3 28
#> 29  95.50729 -2.24857604  3 29
#> 30  89.94790 -1.30668702  3 30
#> 31  94.93657  0.96292158  4 31
#> 32  89.43308 -0.26999602  4 32
#> 33  93.95513  0.70468768  4 33
#> 34 108.92262 -0.94204826  4 34
#> 35 111.87577 -1.25829569  4 35
#> 36  93.58257  0.76671910  4 36
#> 37 102.60717 -0.03750549  4 37
#> 38 117.37950 -0.04034886  4 38
#> 39  79.54187  1.40282663  4 39
#> 40  96.65813  0.49841992  4 40
#> 
#
#Non-normal data:
#
#a) Log-normal data
#
data_gen(Ni=rep(10,4), D=4, M=2 ,ty =NULL, k=c(7,-2), b1=1,
         x1=rnorm(40, 0,1), b2 = 0.5, x2 = rnorm(40, 0,1),
         b3 = NULL, x3 = NULL, b4 = NULL, x4 = NULL,
         xh = NULL, Dis=rLOGNO2,l=c(exp, exp),
         sigma = c(0.4, 0.3), sigmah = NULL,  sigmae = NULL,
         costh = NULL, seed = 1234)
#> Generating population 1 of 2
#> Generating population 2 of 2
#> [[1]]
#>            y          x1          x2 sa id
#> 1   153.1009 -2.03442235 -0.28269260  1  1
#> 2   368.6687 -1.37074010  1.23371790  1  2
#> 3  2297.5496  0.55137001  1.01866568  1  3
#> 4  1036.2391 -0.09649603 -1.02542484  1  4
#> 5   413.2503 -0.85160646 -1.05758010  1  5
#> 6   279.6760 -1.26399413 -0.85073518  1  6
#> 7  1001.8833 -0.01982771  1.01480683  1  7
#> 8  3206.1357  0.69137722  0.50493089  1  8
#> 9  4977.5494  1.54652900 -2.03551224  1  9
#> 10  698.1641 -0.48683301  0.12602989  1 10
#> 11  453.9219 -0.55480193  0.77482934  2 11
#> 12 1245.0747  0.29641564  1.70161679  2 12
#> 13  272.2097 -0.86589657  0.32415216  2 13
#> 14 3261.5761  1.30012764  0.70217982  2 14
#> 15  135.5137 -1.65039921  0.25020827  2 15
#> 16  721.3894 -0.12666103 -0.36141974  2 16
#> 17  759.5242 -0.22059238  0.46509142  2 17
#> 18  166.6673 -1.53135803 -1.36546313  2 18
#> 19  421.3858 -0.44632673 -0.24434471  2 19
#> 20 2765.7068  1.41750422 -0.64939877  2 20
#> 21 1044.2789  0.27217848 -1.05082935  3 21
#> 22  399.4286 -0.56541682 -0.09882419  3 22
#> 23 2996.0265  1.23456509 -0.95882680  3 23
#> 24  167.9394 -1.40256835  0.46474583  3 24
#> 25 2566.8056  1.24984026  1.91042895  3 25
#> 26  200.6785 -1.39518170 -0.48135538  3 26
#> 27  269.1320 -0.88589751 -0.73443686  3 27
#> 28  740.9073  0.02238611  0.93434805  3 28
#> 29 2171.2528  0.88871175 -0.50172245  3 29
#> 30 1437.2864  0.59529254 -0.31140960  3 30
#> 31 2285.5114  0.42770972 -0.17781149  4 31
#> 32  363.0060 -1.24265497  1.02243145  4 32
#> 33 1986.5506  0.01154084  0.32791387  4 33
#> 34 2216.2526  0.32756801 -0.40790079  4 34
#> 35 2740.7155  0.61193799 -0.74081023  4 35
#> 36 2127.3023  0.61426577  1.07062929  4 36
#> 37  648.8014 -0.62829627  0.90964417  4 37
#> 38 2236.2411  0.54983075  0.55396953  4 38
#> 39 2301.8096  0.51084087  0.39247560  4 39
#> 40 6308.5860  1.56149141 -0.16103303  4 40
#> 
#> [[2]]
#>              y          x1          x2 sa id
#> 1     91.88083 -2.03442235 -0.28269260  1  1
#> 2    187.41823 -1.37074010  1.23371790  1  2
#> 3   1629.78817  0.55137001  1.01866568  1  3
#> 4    711.19916 -0.09649603 -1.02542484  1  4
#> 5    324.41285 -0.85160646 -1.05758010  1  5
#> 6    264.86366 -1.26399413 -0.85073518  1  6
#> 7    677.26081 -0.01982771  1.01480683  1  7
#> 8   2043.74313  0.69137722  0.50493089  1  8
#> 9   3676.58533  1.54652900 -2.03551224  1  9
#> 10   638.82187 -0.48683301  0.12602989  1 10
#> 11  1449.50612 -0.55480193  0.77482934  2 11
#> 12  2424.12198  0.29641564  1.70161679  2 12
#> 13   891.94300 -0.86589657  0.32415216  2 13
#> 14  8592.99345  1.30012764  0.70217982  2 14
#> 15   263.81169 -1.65039921  0.25020827  2 15
#> 16  1323.28786 -0.12666103 -0.36141974  2 16
#> 17  2369.37850 -0.22059238  0.46509142  2 17
#> 18   538.68264 -1.53135803 -1.36546313  2 18
#> 19  1233.76339 -0.44632673 -0.24434471  2 19
#> 20  7205.09509  1.41750422 -0.64939877  2 20
#> 21  2297.48733  0.27217848 -1.05082935  3 21
#> 22  1029.84108 -0.56541682 -0.09882419  3 22
#> 23  7421.71899  1.23456509 -0.95882680  3 23
#> 24   362.94967 -1.40256835  0.46474583  3 24
#> 25 10200.83150  1.24984026  1.91042895  3 25
#> 26   566.77232 -1.39518170 -0.48135538  3 26
#> 27   734.60503 -0.88589751 -0.73443686  3 27
#> 28  2432.37783  0.02238611  0.93434805  3 28
#> 29  4777.12166  0.88871175 -0.50172245  3 29
#> 30  2901.72994  0.59529254 -0.31140960  3 30
#> 31  1576.84537  0.42770972 -0.17781149  4 31
#> 32   322.22849 -1.24265497  1.02243145  4 32
#> 33  1143.15458  0.01154084  0.32791387  4 33
#> 34  1613.79055  0.32756801 -0.40790079  4 34
#> 35  2175.91866  0.61193799 -0.74081023  4 35
#> 36  2014.13451  0.61426577  1.07062929  4 36
#> 37   565.35011 -0.62829627  0.90964417  4 37
#> 38  1917.90311  0.54983075  0.55396953  4 38
#> 39  1751.49750  0.51084087  0.39247560  4 39
#> 40  4845.68636  1.56149141 -0.16103303  4 40
#> 
#
#b) Dagum data
#
data_gen(Ni=rep(10,4), D=4, M=2 ,ty =NULL,
         k=c(3, log(3.4), -0.4, log(1)), b1=1.5,
         x1=rnorm(40, 0,1), b2 = 0, x2 =NULL,
         b3 = 0.1, x3 = rnorm(40, 0,1), b4 = 0, x4 = NULL,
         xh = NULL, Dis=rGB2, l=c(exp, exp, exp ,exp),
         sigma = c(0.15,0, 0, 0), sigmah = NULL, sigmae = NULL,
         costh = NULL, seed = 1234)
#> Generating population 1 of 2
#> Generating population 2 of 2
#> [[1]]
#>              y          x1           x3 sa id
#> 1    9.3370387 -0.60580789 -0.364673657  1  1
#> 2   27.2052114 -0.03635736  0.886898485  1  2
#> 3    5.2864365 -0.96226612  0.644851839  1  3
#> 4   12.3250964 -0.08921117  1.198554559  1  4
#> 5   14.2397797 -0.51021545 -0.433041996  1  5
#> 6   19.1788452 -0.52933820 -1.138453429  1  6
#> 7    9.1746346 -0.43265190  1.022368252  1  7
#> 8    2.5203522 -1.87319842 -1.359993678  1  8
#> 9   94.8802645  0.94644298 -1.041188066  1  9
#> 10 219.0607546  1.41714774  0.770658014  1 10
#> 11  23.5851687  0.20600423  0.095729705  2 11
#> 12   5.4640376 -0.63712732  0.866324343  2 12
#> 13   2.4869764 -1.28997487  0.814754762  2 13
#> 14   7.8723265 -0.38483061  0.664855416  2 14
#> 15  50.4444427  0.18919051  0.418100137  2 15
#> 16   2.0246491 -1.08430576 -1.592933782  2 16
#> 17   0.2604757 -2.47930386 -1.523679088  2 17
#> 18  12.4913318 -0.41490511  0.037044364  2 18
#> 19   7.0290526 -0.78390394 -1.368341926  2 19
#> 20   6.6465777 -0.64808540 -0.696248344  2 20
#> 21   4.6486653 -0.67988542 -0.614409320  3 21
#> 22  45.5768445 -0.11643134 -0.325792468  3 22
#> 23  56.8543924  0.23572553  1.827295913  3 23
#> 24  39.5192684 -0.09542670 -1.310302854  3 24
#> 25  16.1043053  0.47585138 -1.477946238  3 25
#> 26  12.9711912 -0.27100226  1.178053051  3 26
#> 27  14.6024503 -0.33478788  1.465676229  3 27
#> 28  44.4730807  0.43371039  0.235974238  3 28
#> 29 129.5998722  0.95131856  0.005247909  3 29
#> 30 133.0591306  1.82856117  0.022226421  3 30
#> 31   3.6514279 -0.68364694 -0.367454407  4 31
#> 32  13.2997593 -0.12129654 -0.522356996  4 32
#> 33   8.7562789 -0.31070032  1.860756278  4 33
#> 34   9.1742195 -0.43458354  0.245579498  4 34
#> 35 134.4925697  1.38664126 -0.831685963  4 35
#> 36   1.5204425 -1.77015446 -1.462122376  4 36
#> 37   5.8197409 -0.71850268 -1.976466164  4 37
#> 38  20.1914917 -0.11953433  2.164447236  4 38
#> 39  24.7925343  0.22153580 -1.566138455  4 39
#> 40 499.4556261  2.59726108 -0.645868168  4 40
#> 
#> [[2]]
#>               y          x1           x3 sa id
#> 1     4.7142135 -0.60580789 -0.364673657  1  1
#> 2    15.1275911 -0.03635736  0.886898485  1  2
#> 3     5.0862743 -0.96226612  0.644851839  1  3
#> 4    16.4985241 -0.08921117  1.198554559  1  4
#> 5     9.9082394 -0.51021545 -0.433041996  1  5
#> 6    12.6239028 -0.52933820 -1.138453429  1  6
#> 7     2.6970302 -0.43265190  1.022368252  1  7
#> 8     1.2105933 -1.87319842 -1.359993678  1  8
#> 9    61.0126050  0.94644298 -1.041188066  1  9
#> 10  122.5877269  1.41714774  0.770658014  1 10
#> 11   17.7665125  0.20600423  0.095729705  2 11
#> 12    7.8753132 -0.63712732  0.866324343  2 12
#> 13    4.2992012 -1.28997487  0.814754762  2 13
#> 14   15.8191618 -0.38483061  0.664855416  2 14
#> 15   33.1458710  0.18919051  0.418100137  2 15
#> 16    5.9830657 -1.08430576 -1.592933782  2 16
#> 17    0.3897992 -2.47930386 -1.523679088  2 17
#> 18   45.4531713 -0.41490511  0.037044364  2 18
#> 19    2.9782123 -0.78390394 -1.368341926  2 19
#> 20    9.0186967 -0.64808540 -0.696248344  2 20
#> 21    9.8002090 -0.67988542 -0.614409320  3 21
#> 22   10.8466410 -0.11643134 -0.325792468  3 22
#> 23   18.5640020  0.23572553  1.827295913  3 23
#> 24   30.3098996 -0.09542670 -1.310302854  3 24
#> 25   45.6852857  0.47585138 -1.477946238  3 25
#> 26   13.1708248 -0.27100226  1.178053051  3 26
#> 27   14.0167463 -0.33478788  1.465676229  3 27
#> 28  101.5563907  0.43371039  0.235974238  3 28
#> 29  107.9957841  0.95131856  0.005247909  3 29
#> 30  300.8830595  1.82856117  0.022226421  3 30
#> 31    5.6112933 -0.68364694 -0.367454407  4 31
#> 32   15.8854122 -0.12129654 -0.522356996  4 32
#> 33   16.9468957 -0.31070032  1.860756278  4 33
#> 34    5.0255079 -0.43458354  0.245579498  4 34
#> 35  160.0244288  1.38664126 -0.831685963  4 35
#> 36    2.0045953 -1.77015446 -1.462122376  4 36
#> 37    9.2618902 -0.71850268 -1.976466164  4 37
#> 38   19.8810140 -0.11953433  2.164447236  4 38
#> 39  330.8258241  0.22153580 -1.566138455  4 39
#> 40 1440.7053328  2.59726108 -0.645868168  4 40
#>