Monte Carlo estimation of mean and HCR based on SAE GAMLSS

A function to estimate the mean or/and the HCR based on unit-level small area estimation based on generalized additive models for location, scale and shape

est_saegamlss(
  sample,
  nonsample,
  D,
  Ni,
  ni,
  f1,
  f2 = NULL,
  f3 = NULL,
  f4 = NULL,
  fdis,
  nRS = NULL,
  nCG = NULL,
  R = NULL,
  Dis,
  np,
  param = NULL,
  seed = NULL,
  tau.fix = NULL,
  nu.fix = NULL,
  z = NULL
)

Arguments

sample

A dataset of sampled units. With dependent variable (named y), covariates and small area (named sa)

nonsample

A dataset of non-sampled units. With covariates and small area (named sa)

D

Number of Areas

Ni

1xD vector containing the number of units (in population) for each area

ni

1xD vector containing the number of sampled-units for each area

f1

A formula object for fitting a model to the mu parameter, e.g.f1=y~x+random(sa)

f2

A formula object for fitting a model to the sigma parameter, e.g.f2=y~x+random(sa)

f3

A formula object for fitting a model to the nu parameter, e.g.f3=y~x+random(sa)

f4

A formula object for fitting a model to the tau parameter, e.g.f4=y~x+random(sa)

fdis

The distribution family of the GAMLSS object

nRS

Number of loop to do with the RS() algorithm

nCG

Number of loop to do with the CG() algorithm

R

Number of Monte-Carlo repetition. Default is 50

Dis

Type of distribution in form of rDis where Dis is one of the distribution allowed by GAMLSS

np

Number of parameters of the distribution. i.e. for the normal distribution np=2, for the GB2 distribution np=4

param

The parameter to estimate, "mean" or "HCR". Default is both.

seed

The seed

tau.fix

A value to be fixed to 1 to obtain the Singh-Maddala distribution. Dist must be GB2 and np=4

nu.fix

A value to be fixed to 1 to obtain the Dagum distribution. Dist must be GB2 and np=4

z

The Poverty line. Default is equal to 0.6 of the dependent variable

Value

a list of two list. The first list, named est has:

ME 1xD vector with the estimate of the mean and

HCR 1xD vector with the estimate of the HCR.

The second, named input_var has:

fit a GAMLSS object with all the components of the regression

f1 a formula object used for fitting a model to the mu parameter

f2 a formula object used for fitting a model to the sigma parameter

f3 a formula object used for fitting a model to the nu parameter

f4 a formula object used for fitting a model to the tau parameter

np the number of the parameters of the choose distribution

nRS the number of loop used within the RS() algorithm. Default is 150.

nCG the number of loop used within the CG() algorithm. Default is 150.

R the number of loop used to obtain the monte-carlo estimation

D the number of area

Ni 1xD vector containing the number of units (in population) for each area

ni 1xD vector containing the number of sampled-units for each area

originaldata The dataset of sampled-units

fids The distribution used for the estimation

z The poverty line

Note

With "object"$input_var$fit is possible to use all the classical function used by gamlss

The Small Area have to be denoted with a number from 1 to D

References

Mori, L., & Ferrante, M. R. (2023). Small area estimation under unit-level generalized additive models for location, scale and shape. arXiv e-prints, arXiv-2302.

Graf, M., Marin, J. M., & Molina, I. (2019). A generalized mixed model for skewed distributions applied to small area estimation. Test, 28(2), 565–597.

Author

Lorenzo Mori and Maria Rosaria Ferrante

Examples

#Generate data
#
dep.y=data_gen(Ni=rep(10,4), D=4, M=2 ,ty ="no", k=4, b1=100,
               x1=rnorm(40, 0,1), b2 = NULL, x2 = NULL, b3 = NULL,
               b4 = NULL, x4 = NULL, xh = NULL, Dis=NO,
               l=c(identity), sigma = 6, sigmah = NULL,
               sigmae = 22, costh = NULL, seed = 1234)
#> Generating population 1 of 2
#> Generating population 2 of 2

data=dep.y[[1]]
#
#sample data with a sample fraction of 0.1
#
library(splitstackshape)
#sample data
#
sample=stratified(data, "sa", size=0.1)
#nonsample data
#
nonsample=nonsample(data=data, sample=sample, id=id)
#estimate
est_saegamlss(sample=sample, nonsample=nonsample,
              D=4, Ni=rep(10,4), ni=rep(1,4),
              f1=y~x1+random(sa), f2=NULL, f3=NULL,
              f4=NULL, fdis=NO, nRS=150, nCG=150, R=200,
              Dis=rNO, np=2, param=NULL,
              seed=1234, tau.fix=NULL, nu.fix=NULL)
#> new prediction 
#> New way of prediction in random()  (starting from GAMLSS version 5.0-6) 
#> new prediction 
#> New way of prediction in random()  (starting from GAMLSS version 5.0-6) 
#> Processing estimation loop 1 of 200
#> Processing estimation loop 2 of 200
#> Processing estimation loop 3 of 200
#> Processing estimation loop 4 of 200
#> Processing estimation loop 5 of 200
#> Processing estimation loop 6 of 200
#> Processing estimation loop 7 of 200
#> Processing estimation loop 8 of 200
#> Processing estimation loop 9 of 200
#> Processing estimation loop 10 of 200
#> Processing estimation loop 11 of 200
#> Processing estimation loop 12 of 200
#> Processing estimation loop 13 of 200
#> Processing estimation loop 14 of 200
#> Processing estimation loop 15 of 200
#> Processing estimation loop 16 of 200
#> Processing estimation loop 17 of 200
#> Processing estimation loop 18 of 200
#> Processing estimation loop 19 of 200
#> Processing estimation loop 20 of 200
#> Processing estimation loop 21 of 200
#> Processing estimation loop 22 of 200
#> Processing estimation loop 23 of 200
#> Processing estimation loop 24 of 200
#> Processing estimation loop 25 of 200
#> Processing estimation loop 26 of 200
#> Processing estimation loop 27 of 200
#> Processing estimation loop 28 of 200
#> Processing estimation loop 29 of 200
#> Processing estimation loop 30 of 200
#> Processing estimation loop 31 of 200
#> Processing estimation loop 32 of 200
#> Processing estimation loop 33 of 200
#> Processing estimation loop 34 of 200
#> Processing estimation loop 35 of 200
#> Processing estimation loop 36 of 200
#> Processing estimation loop 37 of 200
#> Processing estimation loop 38 of 200
#> Processing estimation loop 39 of 200
#> Processing estimation loop 40 of 200
#> Processing estimation loop 41 of 200
#> Processing estimation loop 42 of 200
#> Processing estimation loop 43 of 200
#> Processing estimation loop 44 of 200
#> Processing estimation loop 45 of 200
#> Processing estimation loop 46 of 200
#> Processing estimation loop 47 of 200
#> Processing estimation loop 48 of 200
#> Processing estimation loop 49 of 200
#> Processing estimation loop 50 of 200
#> Processing estimation loop 51 of 200
#> Processing estimation loop 52 of 200
#> Processing estimation loop 53 of 200
#> Processing estimation loop 54 of 200
#> Processing estimation loop 55 of 200
#> Processing estimation loop 56 of 200
#> Processing estimation loop 57 of 200
#> Processing estimation loop 58 of 200
#> Processing estimation loop 59 of 200
#> Processing estimation loop 60 of 200
#> Processing estimation loop 61 of 200
#> Processing estimation loop 62 of 200
#> Processing estimation loop 63 of 200
#> Processing estimation loop 64 of 200
#> Processing estimation loop 65 of 200
#> Processing estimation loop 66 of 200
#> Processing estimation loop 67 of 200
#> Processing estimation loop 68 of 200
#> Processing estimation loop 69 of 200
#> Processing estimation loop 70 of 200
#> Processing estimation loop 71 of 200
#> Processing estimation loop 72 of 200
#> Processing estimation loop 73 of 200
#> Processing estimation loop 74 of 200
#> Processing estimation loop 75 of 200
#> Processing estimation loop 76 of 200
#> Processing estimation loop 77 of 200
#> Processing estimation loop 78 of 200
#> Processing estimation loop 79 of 200
#> Processing estimation loop 80 of 200
#> Processing estimation loop 81 of 200
#> Processing estimation loop 82 of 200
#> Processing estimation loop 83 of 200
#> Processing estimation loop 84 of 200
#> Processing estimation loop 85 of 200
#> Processing estimation loop 86 of 200
#> Processing estimation loop 87 of 200
#> Processing estimation loop 88 of 200
#> Processing estimation loop 89 of 200
#> Processing estimation loop 90 of 200
#> Processing estimation loop 91 of 200
#> Processing estimation loop 92 of 200
#> Processing estimation loop 93 of 200
#> Processing estimation loop 94 of 200
#> Processing estimation loop 95 of 200
#> Processing estimation loop 96 of 200
#> Processing estimation loop 97 of 200
#> Processing estimation loop 98 of 200
#> Processing estimation loop 99 of 200
#> Processing estimation loop 100 of 200
#> Processing estimation loop 101 of 200
#> Processing estimation loop 102 of 200
#> Processing estimation loop 103 of 200
#> Processing estimation loop 104 of 200
#> Processing estimation loop 105 of 200
#> Processing estimation loop 106 of 200
#> Processing estimation loop 107 of 200
#> Processing estimation loop 108 of 200
#> Processing estimation loop 109 of 200
#> Processing estimation loop 110 of 200
#> Processing estimation loop 111 of 200
#> Processing estimation loop 112 of 200
#> Processing estimation loop 113 of 200
#> Processing estimation loop 114 of 200
#> Processing estimation loop 115 of 200
#> Processing estimation loop 116 of 200
#> Processing estimation loop 117 of 200
#> Processing estimation loop 118 of 200
#> Processing estimation loop 119 of 200
#> Processing estimation loop 120 of 200
#> Processing estimation loop 121 of 200
#> Processing estimation loop 122 of 200
#> Processing estimation loop 123 of 200
#> Processing estimation loop 124 of 200
#> Processing estimation loop 125 of 200
#> Processing estimation loop 126 of 200
#> Processing estimation loop 127 of 200
#> Processing estimation loop 128 of 200
#> Processing estimation loop 129 of 200
#> Processing estimation loop 130 of 200
#> Processing estimation loop 131 of 200
#> Processing estimation loop 132 of 200
#> Processing estimation loop 133 of 200
#> Processing estimation loop 134 of 200
#> Processing estimation loop 135 of 200
#> Processing estimation loop 136 of 200
#> Processing estimation loop 137 of 200
#> Processing estimation loop 138 of 200
#> Processing estimation loop 139 of 200
#> Processing estimation loop 140 of 200
#> Processing estimation loop 141 of 200
#> Processing estimation loop 142 of 200
#> Processing estimation loop 143 of 200
#> Processing estimation loop 144 of 200
#> Processing estimation loop 145 of 200
#> Processing estimation loop 146 of 200
#> Processing estimation loop 147 of 200
#> Processing estimation loop 148 of 200
#> Processing estimation loop 149 of 200
#> Processing estimation loop 150 of 200
#> Processing estimation loop 151 of 200
#> Processing estimation loop 152 of 200
#> Processing estimation loop 153 of 200
#> Processing estimation loop 154 of 200
#> Processing estimation loop 155 of 200
#> Processing estimation loop 156 of 200
#> Processing estimation loop 157 of 200
#> Processing estimation loop 158 of 200
#> Processing estimation loop 159 of 200
#> Processing estimation loop 160 of 200
#> Processing estimation loop 161 of 200
#> Processing estimation loop 162 of 200
#> Processing estimation loop 163 of 200
#> Processing estimation loop 164 of 200
#> Processing estimation loop 165 of 200
#> Processing estimation loop 166 of 200
#> Processing estimation loop 167 of 200
#> Processing estimation loop 168 of 200
#> Processing estimation loop 169 of 200
#> Processing estimation loop 170 of 200
#> Processing estimation loop 171 of 200
#> Processing estimation loop 172 of 200
#> Processing estimation loop 173 of 200
#> Processing estimation loop 174 of 200
#> Processing estimation loop 175 of 200
#> Processing estimation loop 176 of 200
#> Processing estimation loop 177 of 200
#> Processing estimation loop 178 of 200
#> Processing estimation loop 179 of 200
#> Processing estimation loop 180 of 200
#> Processing estimation loop 181 of 200
#> Processing estimation loop 182 of 200
#> Processing estimation loop 183 of 200
#> Processing estimation loop 184 of 200
#> Processing estimation loop 185 of 200
#> Processing estimation loop 186 of 200
#> Processing estimation loop 187 of 200
#> Processing estimation loop 188 of 200
#> Processing estimation loop 189 of 200
#> Processing estimation loop 190 of 200
#> Processing estimation loop 191 of 200
#> Processing estimation loop 192 of 200
#> Processing estimation loop 193 of 200
#> Processing estimation loop 194 of 200
#> Processing estimation loop 195 of 200
#> Processing estimation loop 196 of 200
#> Processing estimation loop 197 of 200
#> Processing estimation loop 198 of 200
#> Processing estimation loop 199 of 200
#> Processing estimation loop 200 of 200
#> $est
#> $est$ME
#> [1]  54.31083 -26.55182 -89.80459  63.80178
#> 
#> $est$HCR
#> [1] 0.2385 0.5570 0.8695 0.2820
#> 
#> 
#> $input_var
#> $input_var$fit
#> 
#> Family:  c("NO", "Normal") 
#> Fitting method: mixed(substitute(nRS), substitute(nRG)) 
#> 
#> Call:  gamlss::gamlss(formula = f1, sigma.formula = f2, nu.formula = f3,  
#>     tau.formula = f4, family = substitute(fdis), data = sample,  
#>     method = mixed(substitute(nRS), substitute(nRG)),      trace = F) 
#> 
#> Mu Coefficients:
#> (Intercept)           x1   random(sa)  
#>       18.63       116.45           NA  
#> Sigma Coefficients:
#> (Intercept)  
#>        3.94  
#> 
#>  Degrees of Freedom for the fit: 3 Residual Deg. of Freedom   0.9998 
#> Global Deviance:     42.875 
#>             AIC:     48.8753 
#>             SBC:     47.0341 
#> 
#> $input_var$f1
#> y ~ x1 + random(sa)
#> <environment: 0x0000020378c291c0>
#> 
#> $input_var$f2
#> y ~ 1
#> <environment: 0x0000020379c3de08>
#> 
#> $input_var$f3
#> y ~ 1
#> <environment: 0x0000020379c3de08>
#> 
#> $input_var$f4
#> y ~ 1
#> <environment: 0x0000020379c3de08>
#> 
#> $input_var$fdis
#> function (mu.link = "identity", sigma.link = "log") 
#> {
#>     mstats <- checklink("mu.link", "Normal", substitute(mu.link), 
#>         c("inverse", "log", "identity", "own"))
#>     dstats <- checklink("sigma.link", "Normal", substitute(sigma.link), 
#>         c("inverse", "log", "identity", "own"))
#>     structure(list(family = c("NO", "Normal"), parameters = list(mu = TRUE, 
#>         sigma = TRUE), nopar = 2, type = "Continuous", mu.link = as.character(substitute(mu.link)), 
#>         sigma.link = as.character(substitute(sigma.link)), mu.linkfun = mstats$linkfun, 
#>         sigma.linkfun = dstats$linkfun, mu.linkinv = mstats$linkinv, 
#>         sigma.linkinv = dstats$linkinv, mu.dr = mstats$mu.eta, 
#>         sigma.dr = dstats$mu.eta, dldm = function(y, mu, sigma) (1/sigma^2) * 
#>             (y - mu), d2ldm2 = function(sigma) -(1/sigma^2), 
#>         dldd = function(y, mu, sigma) ((y - mu)^2 - sigma^2)/(sigma^3), 
#>         d2ldd2 = function(sigma) -(2/(sigma^2)), d2ldmdd = function(y) rep(0, 
#>             length(y)), G.dev.incr = function(y, mu, sigma, ...) -2 * 
#>             dNO(y, mu, sigma, log = TRUE), rqres = expression(rqres(pfun = "pNO", 
#>             type = "Continuous", y = y, mu = mu, sigma = sigma)), 
#>         mu.initial = expression({
#>             mu <- (y + mean(y))/2
#>         }), sigma.initial = expression({
#>             sigma <- rep(sd(y), length(y))
#>         }), mu.valid = function(mu) TRUE, sigma.valid = function(sigma) all(sigma > 
#>             0), y.valid = function(y) TRUE, mean = function(mu, 
#>             sigma) mu, variance = function(mu, sigma) sigma^2), 
#>         class = c("gamlss.family", "family"))
#> }
#> <bytecode: 0x0000020379ece788>
#> <environment: namespace:gamlss.dist>
#> 
#> $input_var$np
#> [1] 2
#> 
#> $input_var$nRS
#> [1] 150
#> 
#> $input_var$nCG
#> [1] 150
#> 
#> $input_var$R
#> [1] 200
#> 
#> $input_var$D
#> [1] 4
#> 
#> $input_var$Ni
#> [1] 10 10 10 10
#> 
#> $input_var$ni
#> [1] 1 1 1 1
#> 
#> $input_var$nu.fix
#> NULL
#> 
#> $input_var$tau.fix
#> NULL
#> 
#> $input_var$param
#> [1] "both"
#> 
#> $input_var$origindata
#>             y         x1 sa id
#> 1:   74.79072  0.5662160  1  1
#> 2:   14.21888  0.5220576  2 13
#> 3: -113.16150 -1.0982089  3 28
#> 4:  149.60913  0.4473357  4 36
#> 
#> $input_var$z
#> [1] 2.524085
#> 
#>