Bootstrap Mean Square Error for SAE GAMLSS

Compute the bootstrap MSE for the estimation of the mean and the HCR

mse_saegamlss(
  est,
  D,
  Ni,
  loop = NULL,
  l,
  Dis,
  Iden = F,
  samplesize,
  data,
  cov1,
  cov2 = NULL,
  cov3 = NULL,
  cov4 = NULL,
  seed = NULL
)

Arguments

est

An object obtained with est_saegamlss()

D

Number of Areas

Ni

1xD vector containing the number of units (in population) for each area

loop

Number of loop of bootstrap. Default is 200

l

A vector, of maximum length equal to 4, in which each space is a function (the inverse of the link-function) to be applied to the corresponding parameter, i.e. mu, sigma, nu and tau

Dis

Type of distribution in form of rDis where Dis is one of the distribution allowe by GAMLSS

Iden

TRUE if the Identity link function is used for mu. Default is FALSE

samplesize

The desired sample size. If size is a value between 0 and 1 expressed as a decimal, size is set to be proportional to the number of observations per group. If size is a single positive integer, it will be assumed that you want the same number of samples from each group. If size is a named vector, the function will check to see whether the length of the vector matches the number of groups and that the names match the group names

data

A dataset. Mandatory columns: covariates and areas (named sa). Additional columns are admitted.

cov1

A matrix or a data frame with covariates for the whole population used for mu. If an intercept is used the first columns have to be a vector of 1

cov2

A matrix or a data frame with covariates for the whole population used for sigma. If an intercept is used the first columns have to be a vector of 1

cov3

A matrix or a data frame with covariates for the whole population used for nu. If an intercept is used the first columns have to be a vector of 1

cov4

A matrix or a data frame with covariates for the whole population used for tau. If an intercept is used the first columns have to be a vector of 1

seed

The seed

Value

MSE_mean A vector with the MSE of the mean for each area

MSE_HCR A vector with the MSE of the HCR for each area

All the value returned by est_saegamlss().

References

Mori, L., & Ferrante, M. R. (2023). Small area estimation under unit-level generalized additive models for location, scale and shape. arXiv e-prints, arXiv-2302.

Graf, M., Marin, J. M., & Molina, I. (2019). A generalized mixed model for skewed distributions applied to small area estimation. Test, 28(2), 565–597.

Author

Lorenzo Mori and Maria Rosaria Ferrante

Examples

dep.y=data_gen(Ni=rep(10,4), D=4, M=2 ,ty ="no", k=4, b1=100,
               x1=rnorm(40, 0,1), b2 = NULL, x2 = NULL, b3 = NULL,
               x3 = NULL,  b4 = NULL, x4 = NULL, xh = NULL,
               Dis, l=c(identity), sigma = 6, sigmah = NULL,
               sigmae = 22,  costh = NULL, seed = 1234)
#> Generating population 1 of 2
#> Generating population 2 of 2
data=dep.y[[1]]
#
#sample data with a sample fraction of 0.5
#
library(splitstackshape)
#sample data
#
sample=stratified(data, "sa", size=0.5)
#nonsample data
#
nonsample=nonsample(data=data, sample=sample, id=id)
#estimate
est=est_saegamlss(sample=sample, nonsample=nonsample, D=4,
                 Ni=rep(10,4), ni=rep(1,4), f1=y~x1+random(sa),
                 f2=NULL, f3=NULL, f4=NULL, fdis=NO, nRS=150,
                 nCG=150, R=2, Dis=rNO, np=2, param=NULL,
                 seed=1234, tau.fix=NULL, nu.fix=NULL)
#> new prediction 
#> New way of prediction in random()  (starting from GAMLSS version 5.0-6) 
#> new prediction 
#> New way of prediction in random()  (starting from GAMLSS version 5.0-6) 
#> Processing estimation loop 1 of 2
#> Processing estimation loop 2 of 2
#
#covariates
#
x=data.frame(rep(1, nrow(data)), "x1"=data$x1)
#
#compute the MSE
#
MSE=mse_saegamlss(est=est, D=4, Ni=rep(10,4), loop=2,
               l=c(identity), Dis=rNO, Iden = TRUE,
               samplesize=0.1, data=data, cov1=x, cov2 = NULL, cov3 = NULL,
               cov4 = NULL, seed=1234)
#> new prediction 
#> New way of prediction in random()  (starting from GAMLSS version 5.0-6) 
#> new prediction 
#> New way of prediction in random()  (starting from GAMLSS version 5.0-6) 
#> Processing estimation loop 1 of 2
#> Processing estimation loop 2 of 2
#> Processing mse loop 1 of 2
#> new prediction 
#> New way of prediction in random()  (starting from GAMLSS version 5.0-6) 
#> new prediction 
#> New way of prediction in random()  (starting from GAMLSS version 5.0-6) 
#> Processing estimation loop 1 of 2
#> Processing estimation loop 2 of 2
#> Processing mse loop 2 of 2
MSE$est_mse$MSE_mean
#> [1] 390.8433 832.6707 688.5483 207.8617
MSE$est_mse$MSE_HCR
#> [1] 0.05625 0.00125 0.02000 0.00625